requirements.txt
python-dotenv
telethon
python-telegram-bot==13.15